requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==24.3.0
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.4
django-environ==0.11.2
django-rest-swagger==2.2.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
MarkupSafe==3.0.2
openapi-codec==1.3.2
packaging==24.2
pluggy==1.5.0
prompt_toolkit==3.0.48
psycopg2==2.9.10
pytest==8.3.4
pytest-django==4.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
simplejson==3.19.3
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
